LPS-196226 Transform JSBundleConfigRegistry into JSBundleConfigRegistryUtil and apply to usages
